Fix compiler warnings for VS2010 /W4 in config.c and encmain.c.
- Working towards issue #11.
- Widened datatypes for cfg struct members that take values from atoi to full
  ints so that bounds checking can be done after parsing without overflow.

+++ src/config.h
@@ -36,29 +36,29 @@ typedef struct
   char *input;      /*!< \brief Pointer to input filename  */
   char *output;     /*!< \brief Pointer to output filename */
   char *debug;      /*!< \brief Pointer to debug output    */
-  int8_t qp;        /*!< \brief Quantization parameter */
-  int16_t intra_period; /*!< \brief the period of intra frames in stream */
+  int32_t qp;        /*!< \brief Quantization parameter */
+  int32_t intra_period; /*!< \brief the period of intra frames in stream */
   int32_t frames;  /*!< \brief Number of frames to decode */
   int32_t width;   /*!< \brief frame width */
   int32_t height;  /*!< \brief frame height */
-  int8_t deblock_enable; /*!< \brief Flag to enable deblocking filter */
-  int8_t sao_enable;     /*!< \brief Flag to enable sample adaptive offset filter */
-  int8_t deblock_beta;   /*!< \brief (deblocking) beta offset (div 2), range -6...6 */
-  int8_t deblock_tc;     /*!< \brief (deblocking) tc offset (div 2), range -6...6 */
+  int32_t deblock_enable; /*!< \brief Flag to enable deblocking filter */
+  int32_t sao_enable;     /*!< \brief Flag to enable sample adaptive offset filter */
+  int32_t deblock_beta;   /*!< \brief (deblocking) beta offset (div 2), range -6...6 */
+  int32_t deblock_tc;     /*!< \brief (deblocking) tc offset (div 2), range -6...6 */
   struct
   {
-    int16_t sar_width;   /*!< \brief the horizontal size of the sample aspect ratio (in arbitrary units) */
-    int16_t sar_height;  /*!< \brief the vertical size of the sample aspect ratio (in the same arbitrary units as sar_width). */
+    int32_t sar_width;   /*!< \brief the horizontal size of the sample aspect ratio (in arbitrary units) */
+    int32_t sar_height;  /*!< \brief the vertical size of the sample aspect ratio (in the same arbitrary units as sar_width). */
     int8_t overscan;     /*!< \brief Crop overscan setting */
     int8_t videoformat;  /*!< \brief Video format */
     int8_t fullrange;    /*!< \brief Flag to indicate full-range */
     int8_t colorprim;    /*!< \brief Color primaries */
     int8_t transfer;     /*!< \brief Transfer characteristics */
     int8_t colormatrix;  /*!< \brief Color matrix coefficients */
-    int8_t chroma_loc;   /*!< \brief Chroma sample location */
+    int32_t chroma_loc;   /*!< \brief Chroma sample location */
   } vui;
-  int8_t aud_enable;     /*!< \brief Flag to use access unit delimiters */
-  int8_t ref_frames;     /*!< \brief number of reference frames to use */
+  int32_t aud_enable;     /*!< \brief Flag to use access unit delimiters */
+  int32_t ref_frames;     /*!< \brief number of reference frames to use */
   char * cqmfile;        /*!< \brief Pointer to custom quantization matrices filename */
 } config;
